refactor(common)
